@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "version": "1.9.3",
   "main": "dist/index.js",
   "name": "@stakewise/v3-deposit-data-parser",
@@ -11,9 +12,9 @@
   "module": "dist/index.esm.js",
   "scripts": {
     "prepare": "npm run beforePublish",
-    "test": "jest --clearCache && jest --onlyChanged",
+    "test": "jest --clearCache && jest --all",
     "build": "npm run test && npm run prepare && rm -rf ./dist && npm run rollup",
-    "rollup": "rollup --config rollup.config.js --bundleConfigAsCjs",
+    "rollup": "rollup --config rollup.config.js",
     "beforePublish": "node ./scripts/beforePublish",
     "publish": "npm run build && npm publish --tag=latest"
   },
@@ -34,27 +35,28 @@
     "./package.json": "./package.json"
   },
   "devDependencies": {
-    "@chainsafe/ssz": "0.11.1",
-    "@openzeppelin/merkle-tree": "1.0.5",
-    "@rollup/plugin-commonjs": "25.0.4",
-    "@rollup/plugin-json": "6.0.0",
-    "@rollup/plugin-node-resolve": "15.2.1",
-    "@rollup/plugin-terser": "0.4.3",
-    "@rollup/plugin-typescript": "11.1.3",
-    "@types/jest": "29.5.12",
-    "@types/rollup-plugin-peer-deps-external": "2.2.1",
-    "@typescript-eslint/eslint-plugin": "6.7.0",
-    "@typescript-eslint/parser": "6.7.0",
-    "bls-eth-wasm": "1.1.1",
-    "eslint": "8.56.0",
-    "eslint-plugin-import": "2.28.1",
+    "@chainsafe/ssz": "1.2.0",
+    "@openzeppelin/merkle-tree": "1.0.8",
+    "@rollup/plugin-commonjs": "28.0.3",
+    "@rollup/plugin-json": "6.1.0",
+    "@rollup/plugin-node-resolve": "16.0.1",
+    "@rollup/plugin-terser": "0.4.4",
+    "@rollup/plugin-typescript": "12.1.2",
+    "@types/jest": "29.5.14",
+    "@types/rollup-plugin-peer-deps-external": "2.2.5",
+    "@typescript-eslint/eslint-plugin": "8.33.0",
+    "@typescript-eslint/parser": "8.33.0",
+    "bls-eth-wasm": "1.4.0",
+    "eslint": "9.27.0",
+    "eslint-plugin-import": "2.31.0",
     "jest": "29.7.0",
-    "rollup": "3.29.1",
-    "rollup-plugin-delete": "2.0.0",
-    "rollup-plugin-dts": "6.0.2",
+    "rollup": "4.41.1",
+    "rollup-plugin-delete": "3.0.1",
+    "rollup-plugin-dts": "6.2.1",
     "rollup-plugin-peer-deps-external": "2.2.4",
-    "ts-jest": "29.2.4",
-    "ts-node": "10.9.1",
-    "typescript": "5.1.6"
+    "ts-jest": "29.3.4",
+    "ts-node": "10.9.2",
+    "tslib": "2.8.1",
+    "typescript": "5.8.3"
   }
 }
