@@ -7,7 +7,7 @@ import resolve from '@rollup/plugin-node-resolve'
 import typescript from '@rollup/plugin-typescript'
 import peerDepsExternal from 'rollup-plugin-peer-deps-external'
 
-import pkg from './package.json'
+import pkg from './package.json' with { type: 'json' }
 
 
 const config = [
@@ -29,7 +29,6 @@ const config = [
       },
     ],
     plugins: [
-      // @ts-ignore: this plugin has types for old version of jest :(
       peerDepsExternal(),
       resolve(),
       commonjs({
